Split AccountBalance and AccountController
