+++ src/main/java/com/cooperex/cex/controller/RegisterController.java
@@ -1,6 +1,6 @@
 package com.cooperex.cex.controller;
 import com.cooperex.cex.model.Register;
-import com.cooperex.cex.model.RegisterDAO;
+import com.cooperex.cex.dao.RegisterDAO;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
